implement migrateStreamingStyles utility
When using streaming rendering, style blocks are emitted in chunks directly
next to the HTML they reference. In order to prevent errors during rehydration
(since React doesn't know about the style blocks we are interleaving) this
method relocates all styled-component blocks to the end of `<head>`.

Although the naming of this could be made generic, I thought
it would be a good idea to be specific so people strongly associate
the functionality with the streaming process.