Revert #1224
It was causing a regression in SSR which easily goes unnoticed and
doesn't only affect this PR but broke master too.

In general minification is not necessary because the SSR'd CSS code
has already run though stylis. Together with gzip this means that the
actual gains in size are negligible.

Since I'm not sure how well it plays with extractCompsFromCSS.js in
*all* scenarios I'll revert it (for now)